test: consolidate model and metadata fixtures
